[fix] abstract store structure fix
